test(KEN-1241): isolate session JSONL fixtures
Separate streaming line reads from cached user-message extraction and give each case its own directory with unconditional cleanup.

Validation: focused suite, Pi QoL package, preflight, doc-limits, Pi package policy, and copied controls passed. Specialist review passed. Full repository validation awaits the root-owned slot.
